requirements.txt
pyserial
pyrebase
python-dotenv
smbus